+++ CHANGELOG.md
@@ -13,6 +13,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - run on provided `MachineInfo` by default; bare metal M-mode environment should gate `machine`
 - doc: grammar tweaks in hsm module
+- update dependency `sbi-spec` to v0.0.6, use `Physical` struct from `sbi-spec` crate.
 
 ### Removed